backend/requirements/base.txt
# Django
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-filter==23.5
django-environ==0.11.*

# Database
psycopg2-binary==2.9.*
pymongo==4.6.*

# Caching & Queue
redis==5.0.*
celery==5.3.*
django-redis==5.4.*

# Web Scraping
scrapy==2.11.*
selenium==4.16.*
beautifulsoup4==4.12.*
requests==2.31.*
PySocks==1.7.*
stem==1.8.*

# Security Scanning
python-nmap==0.7.*
dnspython==2.6.*
tldextract==5.1.*

# Domain & Reputation APIs
whois==0.9.*

# Security
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.*
gunicorn==21.2.*
whitenoise==6.6.*
Pillow==10.1.*

# Monitoring
sentry-sdk==1.39.*
psutil==5.9.*
django-celery-beat==2.5.*

# Telegram
python-telegram-bot==20.7.*
requests==2.31.*
